Upgrade openssl to use local openssl
python.props is now directing the python core to the openssl dir
that is built prior to the core for openssl dlls.

Using this as a reference:
https://github.com/reiniert/cpython/commit/cf5efa376809b96afc4f1c774ebe2d3f074a167e

Upgrade from 1.0.2 to 1.1.1:
Remove the libeay and ssleay projects and link to local  openssl
This is specifically using the DLLS with the 1_1 suffix, which
is the suffix for both 64 and 32 bit DLLs we are building.