Upgrade openssl to use openssl-bin-1.1.1j
python.props is now directing the python core to the
openssl-bin dir for openssl dlls.

Using this as a reference:
https://github.com/reiniert/cpython/commit/cf5efa376809b96afc4f1c774ebe2d3f074a167e

Remove the libeay and ssleay projects and use openssl from the
openssl-bin.  This is specifically using the DLLS with the 1_1-x64
suffix, we may need to adjust this in openssl.props for 32 bit builds.

This should be updated to use the existing output of the openssl build.